+++ libminifi/test/unit/FileStreamTests.cpp
@@ -23,6 +23,7 @@
 
 #include "io/FileStream.h"
 #include "../TestBase.h"
+#include "utils/gsl.h"
 
 TEST_CASE("TestFileOverWrite", "[TestFiles]") {
   TestController testController;
@@ -39,7 +40,7 @@ TEST_CASE("TestFileOverWrite", "[TestFiles]") {
 
   minifi::io::FileStream stream(path, 0, true);
   std::vector<uint8_t> readBuffer;
-  REQUIRE(stream.read(readBuffer, stream.size()) == stream.size());
+  REQUIRE(stream.read(readBuffer, gsl::narrow<int>(stream.size())) == stream.size());
 
   uint8_t* data = readBuffer.data();
 
@@ -53,7 +54,7 @@ TEST_CASE("TestFileOverWrite", "[TestFiles]") {
 
   std::vector<uint8_t> verifybuffer;
 
-  REQUIRE(stream.read(verifybuffer, stream.size()) == stream.size());
+  REQUIRE(stream.read(verifybuffer, gsl::narrow<int>(stream.size())) == stream.size());
 
   data = verifybuffer.data();
 
@@ -77,7 +78,7 @@ TEST_CASE("TestFileBadArgumentNoChange", "[TestLoader]") {
 
   minifi::io::FileStream stream(path, 0, true);
   std::vector<uint8_t> readBuffer;
-  REQUIRE(stream.read(readBuffer, stream.size()) == stream.size());
+  REQUIRE(stream.read(readBuffer, gsl::narrow<int>(stream.size())) == stream.size());
 
   uint8_t* data = readBuffer.data();
 
@@ -91,7 +92,7 @@ TEST_CASE("TestFileBadArgumentNoChange", "[TestLoader]") {
 
   std::vector<uint8_t> verifybuffer;
 
-  REQUIRE(stream.read(verifybuffer, stream.size()) == stream.size());
+  REQUIRE(stream.read(verifybuffer, gsl::narrow<int>(stream.size())) == stream.size());
 
   data = verifybuffer.data();
 
@@ -115,7 +116,7 @@ TEST_CASE("TestFileBadArgumentNoChange2", "[TestLoader]") {
 
   minifi::io::FileStream stream(path, 0, true);
   std::vector<uint8_t> readBuffer;
-  REQUIRE(stream.read(readBuffer, stream.size()) == stream.size());
+  REQUIRE(stream.read(readBuffer, gsl::narrow<int>(stream.size())) == stream.size());
 
   uint8_t* data = readBuffer.data();
 
@@ -129,7 +130,7 @@ TEST_CASE("TestFileBadArgumentNoChange2", "[TestLoader]") {
 
   std::vector<uint8_t> verifybuffer;
 
-  REQUIRE(stream.read(verifybuffer, stream.size()) == stream.size());
+  REQUIRE(stream.read(verifybuffer, gsl::narrow<int>(stream.size())) == stream.size());
 
   data = verifybuffer.data();
 
@@ -153,7 +154,7 @@ TEST_CASE("TestFileBadArgumentNoChange3", "[TestLoader]") {
 
   minifi::io::FileStream stream(path, 0, true);
   std::vector<uint8_t> readBuffer;
-  REQUIRE(stream.read(readBuffer, stream.size()) == stream.size());
+  REQUIRE(stream.read(readBuffer, gsl::narrow<int>(stream.size())) == stream.size());
 
   uint8_t* data = readBuffer.data();
 
@@ -167,7 +168,7 @@ TEST_CASE("TestFileBadArgumentNoChange3", "[TestLoader]") {
 
   std::vector<uint8_t> verifybuffer;
 
-  REQUIRE(stream.read(nullptr, stream.size()) == -1);
+  REQUIRE(stream.read(nullptr, gsl::narrow<int>(stream.size())) == -1);
 
   data = verifybuffer.data();
 
@@ -191,7 +192,7 @@ TEST_CASE("TestFileBeyondEnd3", "[TestLoader]") {
 
   minifi::io::FileStream stream(path, 0, true);
   std::vector<uint8_t> readBuffer;
-  REQUIRE(stream.read(readBuffer, stream.size()) == stream.size());
+  REQUIRE(stream.read(readBuffer, gsl::narrow<int>(stream.size())) == stream.size());
 
   uint8_t* data = readBuffer.data();
 
@@ -226,7 +227,7 @@ TEST_CASE("TestFileExceedSize", "[TestLoader]") {
 
   minifi::io::FileStream stream(path, 0, true);
   std::vector<uint8_t> readBuffer;
-  REQUIRE(stream.read(readBuffer, stream.size()) == stream.size());
+  REQUIRE(stream.read(readBuffer, gsl::narrow<int>(stream.size())) == stream.size());
 
   uint8_t* data = readBuffer.data();
