+++ main/AgentDocs.cpp
@@ -47,7 +47,7 @@ std::string AgentDocs::extractClassName(const std::string &processor) const {
   auto lastOfIdx = processor.find_last_of(".");
   if (lastOfIdx != std::string::npos) {
     lastOfIdx++;  // if a value is found, increment to move beyond the .
-    int nameLength = processor.length() - lastOfIdx;
+    size_t nameLength = processor.length() - lastOfIdx;
     std::string processorName = processor.substr(lastOfIdx, nameLength);
     return processorName;
   }
